Move ECR pull through transform test to unit
